Pass third argument to linspace as integer
Ceil returns float, so it requires a further cast to int before linspace accepts it as its third positional argument. Fixes <https://github.com/acolite/acolite/issues/5>

+++ acolite/shared/rsr_convolute_dict.py
@@ -4,12 +4,13 @@
 ## 2016-11
 ## modifications: 2017-10-17 (QV) moved wave range and step to keywords, renamed from interprsr
 ##                2020-03-02 (QV) added ceil to number of elements in linspace
+##                2020-06-02 (SV) pass ceil to int to ensure that linspace gets an integer argument
 
 def rsr_convolute_dict(wave_data, data, rsr, wave_range=[0.2,2.4], wave_step=0.001):
     from numpy import linspace, interp, nan, zeros, ceil
     
     ## set up wavelength space
-    wave_hyper = linspace(wave_range[0],wave_range[1],ceil(((wave_range[1]-wave_range[0])/wave_step)+2))
+    wave_hyper = linspace(wave_range[0],wave_range[1],int(ceil(((wave_range[1]-wave_range[0])/wave_step)+2)))
 
     ## interpolate RSR to same dimensions
     rsr_hyper = dict()
